requirements.txt
tensorflow
opencv-python
numpy
Pillow
requests
